Move types to Registration namespace
