+++ src/main/Yardarm.NewtonsoftJson/JsonCreateDefaultRegistryEnricher.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
-using Yardarm.Enrichment;
+using Yardarm.Enrichment.Registration;
 using static Microsoft.CodeAnalysis.CSharp.SyntaxFactory;
 
 namespace Yardarm.NewtonsoftJson
