+++ src/main/Yardarm.SystemTextJson/JsonCreateDefaultRegistryEnricher.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
-using Yardarm.Enrichment;
+using Yardarm.Enrichment.Registration;
 using Yardarm.SystemTextJson.Internal;
 using static Microsoft.CodeAnalysis.CSharp.SyntaxFactory;
