+++ src/main/Yardarm/Enrichment/Registration/IRegistrationEnricher.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.CodeAnalysis.CSharp.Syntax;
 
-namespace Yardarm.Enrichment;
+namespace Yardarm.Enrichment.Registration;
 
 /// <summary>
 /// An expression enricher used by extensions to register themselves in the client.
